refactor: switch response from decimal to double for stock prices

+++ src/Services/Stock/Stock.Application/Common/Models/StockPriceResultDto.cs
@@ -4,7 +4,7 @@ public record StockPriceResultDto
 {
     public string Id { get; init; } = default!;
     public string Symbol { get; init; } = default!;
-    public decimal? Price { get; init; }
+    public double? Price { get; init; }
     public bool IsActive { get; init; }
     public DateTime? LastPriceUpdate { get; init; }
     public DateTime Created { get; init; }
