+++ src/Services/Stock/Stock.Application/Queries/Stock/GetStockById.cs
@@ -71,7 +71,7 @@ public class GetStockByIdHandler : IRequestHandler<GetStockById, GetStockByIdRes
         return new GetStockByIdResponse
         {
             LastPriceUpdate = item.LastPriceUpdate,
-            Price = item.Price == -1 ? default(decimal?) : item.Price,
+            Price = item.Price == -1 ? default(double?) : (double)item.Price,
             Symbol = item.Symbol,
             IsActive = item.IsActive,
             Id = _sqids.Encode(item.StockId),
