+++ src/Services/Stock/Stock.Application/Queries/Stock/GetStockBySymbol.cs
@@ -65,7 +65,7 @@ public class GetStockBySymbolHandler : IRequestHandler<GetStockBySymbol, GetStoc
         return new GetStockBySymbolResponse
         {
             LastPriceUpdate = item.LastPriceUpdate,
-            Price = item.Price == -1 ? default(decimal?) : item.Price,
+            Price = item.Price == -1 ? default(double?) : (double)item.Price,
             Symbol = item.Symbol,
             IsActive = item.IsActive,
             Id = _sqids.Encode(item.StockId),
